include, linux, refcount: get rid of global reference count
This should remove the need of locking during the update of
reference counts, provided that no new refcount classes are added
while existing ones are updated, and at the cost of a worst-case
O(n) update (linear on the number of refcount classes of the object).
See the comments in the code for more details on the mechanism.

Signed-off-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>